Moved save events to events.services.ts. Added some aditional error handling for the async functions. This will log errors but not stop the inputting of other events into the Db.
