+++ src/services/event.service.ts
@@ -6,7 +6,7 @@
 
 import { ObjectId } from '../../deps.ts';
 import { db } from '../database/connect.ts';
-import { Event } from '../models/event.model.ts';
+import { CompleteEventType, Event } from '../models/event.model.ts';
 
 const events = db.collection<Event>('events');
 
@@ -18,7 +18,80 @@ const getEventById = async (id: string): Promise<Event | null> => {
   return await events.findOne({ _id: new ObjectId(id) });
 };
 
+const databaseIncludes = async (event: Event): Promise<boolean> => {
+  try {
+
+    const exists = await events.findOne({
+    name: event.name,
+    date: event.date,
+    location: event.location,
+    });
+
+    return exists !== null;
+
+  } catch (error: unknown) {
+    if (error instanceof Error) {
+
+      throw new Error(`Event in Db check failed: ${error.message}`);
+    } else {
+
+      throw new Error('Event in Db check failed with an unknown error');
+    }
+  }
+};
+
+const isCompleteEventType = (obj: unknown): obj is CompleteEventType => {
+  return (
+    typeof obj === 'object' &&
+    obj !== null &&
+    Object.values(obj).every((val) => Array.isArray(val))
+  );
+};
+
+const saveEvents = async (input: Event | CompleteEventType) => {
+    if (isCompleteEventType(input)) {
+
+      for (const eventCategory of Object.values(input)) {
+
+        for (const event of eventCategory) {
+
+          try {
+            if (!(await databaseIncludes(event))) {
+              await events.insertOne(event);
+            }
+
+          } catch (error) {
+
+            if (error instanceof Error) {
+
+              console.error(`Event save failed: ${JSON.stringify(event)}, ${error.message}`);
+            } else {
+
+              console.error('Event save failed with an unknown error');
+            }
+          }
+        }
+      }
+    } else {
+      try {
+        if (!(await databaseIncludes(input))) {
+        await events.insertOne(input);
+        }
+
+      } catch (error) {
+        if (error instanceof Error) {
+
+          console.error(`Event save failed: ${JSON.stringify(input)}, ${error.message}`);
+      } else {
+
+        console.error('Event save failed with an unknown error');
+      }
+    }
+  }  
+};
+
 export const eventService = {
   getAllEvents,
   getEventById,
+  saveEvents
 };
